+++ package.json
@@ -14,6 +14,7 @@
     "@vue-flow/controls": "^1.0.6",
     "@vue-flow/core": "^1.20.0",
     "@vue-flow/minimap": "^1.1.1",
+    "@vue-flow/node-resizer": "^1.3.2",
     "core-js": "^3.29.0",
     "roboto-fontface": "*",
     "vue": "^3.2.0",
